% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_set.R
\docType{data}
\name{dummy_set}
\alias{dummy_set}
\title{A simple data set with stock close prices}
\format{
An object of class \code{data.frame} with 1925 rows and 4 columns.
}
\source{
Yahoo Finance
}
\usage{
dummy_set
}
\description{
A data frame with the close prices for AMZN, NVDA and IBM
}
\keyword{datasets}
